requirements.txt
streamlit
streamlit_drawable_canvas
tensorflow
numpy
pillow
matplotlib